Tests for jwt.util
