+++ classes/fullstack/project1/server/package.json
@@ -63,7 +63,7 @@
     "jest": "^29.7.0",
     "lint-staged": "^15.2.2",
     "supertest": "^6.3.4",
-    "ts-jest": "^29.1.2",
+    "ts-jest": "^29.1.3",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
